Implement first version of the package
The first version of the package is a direct translation of the R code
provided by Professor Florencia Leonardi. The benchmark show the R and
the C++ implementations have the same performance. This is probably due
to the fact that, in the inner loop of the C++ implementation, it's
calling the loglikfun R function provided as argument. Because of this,
the interop between C++ and R is likely to be the bottleneck here. The
first attempt to speed the algorithm up is going to be to remove the
implementation the R function call from the code, by providing a direct
C++ implementation.