Replace r_multivariate by native multivariate
To my surprise, the R version of the algorithm runs as fast as the Cpp
implementation, because the slow part of the algorithm was actually the
multivariate function all along. We should probably consider the R
version of the algorithm, as it's more maintainable and as fast as the
Rcpp version.