Refactor na_action segment funciton argument
`na.omit` flag argument was replaced with an `na_action` function
argument that defaults to `na.omit` function. This allows for more
flexibility, allowing the user of the package to use an identity
funciton in case he or she wants to preserve na values, and also allows
to take a different action altogether.