Implement interleave index splitting
Because a lot of the computing done inside the workers involve sub setting
the data using a `foreach(seg_end = indices)` such that the subset
equals `data[, 1:seg_end]`, we have that the subsets tent to be bigger
the higher the `seg_end`.

In the previous implementation, the indices were split sequentially,
such that c(1, 2, 3, 4, 5, 6) would be split in list(c(1, 2, 3), c(4, 5,
6)) for two workers. That would result in worker 1 doing less work than
worker 2, as worker two is assigned higher indices.

By introducing interleave combination, we have the indices are split
using list(c(1, 3, 5), c(2, 4, 6)), such the work is more evenly
distributed between the two.