Remove unecessary .packages argument from foreach
The `foreach` package takes a `.packages` argument which can specify what
packages should be loaded in the workers of the cluster. the `%doPar%`
function automatically infers variables referenced inside the block
scope, and loads the variables in the cluster accordingly. Because all
the functions of the package are defined in the current environment,
they are already passed to the cluster, dismissing the necessity of
specifying `segmentr` as argument to `.packages`