Fix namespace so that only `@export`ed functions are exported
The initial Rcpp project, created from the RStudio wizard, created
a NAMESPACE that wansn't allowing roxygen2 to override it with the
functions tagged with the `@export` keyword.

In order to get around that, the `rcpp_headers.R` file was included
in order to hint roxygen2 what lines have to be included in the final
NAMESPACE so the Rcpp functions keep on working.