Change definition of changepoint
In the previous implementation, changepoint was the last index of a
given segment. In the new definition, it's now the first index of a
given segment.

The motivation behind this was the name changepoint suggests a point
after which a change took place. In the context of this package,
it makes more sense for it to represent the first index of a segment,
therefore the change of defnition.