Change multivariate default na_action function
Data analyzed in segmentr usually has a handful of rows and many more
columns. The default `na.omit` function removes columns that have NA
values, which meant a lot of significant data was removed from the
dataset. Because of this, de default `na_action` was changed to just
remove the columns in which an NA appears.