Use library rather than require notation
This allows the code to fail whenever it tries to load a package that is
not installed. `require` will try to load the package but just but will
just display a warning if it's not there. `library` will stop execution
showing an error message, which is the correct use-case when rendering a
bookdown document.